+++ src/Sanitiser.Core/Sanitisers/ISanitiser.cs
@@ -4,7 +4,7 @@ namespace Umbraco.Community.Sanitiser.sanitisers;
 
 public interface ISanitiser : IDiscoverable
 {
-    public Task Sanitise();
+    public Task Sanitise(SanitisationContext context);
 
     public bool IsEnabled();
 }